1. DB env객체 커넥션 매니저 멤버로 변경 2. playerLogin 추가 3. 유저Session새로 생성

+++ DreamWorld_Server/MsgDispatcher/MsgDispatcher.h
@@ -10,11 +10,11 @@ namespace DreamWorld {
 using MsgHandler = std::function<void(sh::IO_Engine::ISessionPtr, BYTE*)>;
 class MsgDispatcher {
  public:
-  bool AddMsgHandler(const uint16_t msgId, MsgHandler handler);
+  bool AddMsgHandler(const uint8_t msgId, MsgHandler handler);
 
-  bool GetHandler(const uint16_t msgId, MsgHandler& handler);
+  bool GetHandler(const uint8_t msgId, MsgHandler& handler);
 
  private:
-  std::unordered_map<uint16_t, MsgHandler> m_dispatcher;
+  std::unordered_map<uint8_t, MsgHandler> m_dispatcher;
 };
 }  // namespace DreamWorld
